+++ matrix.h
@@ -26,14 +26,14 @@ typedef struct matrix matrix_t;
 int matrix_init(matrix_t *matrix, int precision);
 
 void matrix_set_explicit(matrix_t *matrix,
-			 double m00, double m01, double m02,
-			 double m10, double m11, double m12,
-			 double m20, double m21, double m22);
+			 float m00, float m01, float m02,
+			 float m10, float m11, float m12,
+			 float m20, float m21, float m22);
 void matrix_set_identity(matrix_t *matrix);
-void matrix_set_translate(matrix_t *matrix, double x_off, double y_off);
-void matrix_set_scale(matrix_t *matrix, double x_scale, double y_scale);
-void matrix_set_rotate(matrix_t *matrix, double theta);
-void matrix_set_rotate_about(matrix_t *matrix, double theta, double x, double y);
+void matrix_set_translate(matrix_t *matrix, float x_off, float y_off);
+void matrix_set_scale(matrix_t *matrix, float x_scale, float y_scale);
+void matrix_set_rotate(matrix_t *matrix, float theta);
+void matrix_set_rotate_about(matrix_t *matrix, float theta, float x, float y);
 
 void matrix_copy(matrix_t *dest, matrix_t *src);
 void matrix_multiply(matrix_t *result, matrix_t *a, matrix_t *b);
